api/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
redis==4.6.0
pika==1.3.2
psycopg2-binary==2.9.7
sqlalchemy==2.0.20
alembic==1.12.0
python-dotenv==1.0.0